requirements.txt
streamlit>=1.37
requests>=2.32
beautifulsoup4>=4.12
lxml>=5.2
pandas>=2.2
openpyxl>=3.1
python-dateutil>=2.9
PyYAML>=6.0
# Optional LLM
openai>=1.40
tiktoken>=0.7
